# public/denis23x.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "description": "This API provides access to a collection of books, users, and reviews, allowing clients to perform various operations like viewing available books, creating and managing users, and posting or retrieving book reviews.\n\n- **Books**: Retrieve book details, search for books, and view books.\n- **Users**: Manage users, view user profiles.\n- **Reviews**: Add, view, update, or delete reviews for books.",
    "title": "Books Showcase Demo",
    "contact": {
      "email": "denix23x@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/api/v1/placeholder/books/{id}": {
      "get": {
        "summary": "Get Book",
        "deprecated": false,
        "description": "",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Book"
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": {
                        "id": 343,
                        "title": "curis cilicium",
                        "description": "Veritatis curtus optio carcer adopto esse. Corrupti utrum terminatio cauda ultra uxor territo thermae trado. Deorsum appositus acerbitas somniculosus sumptus iste credo veritas enim.",
                        "cover": "https://picsum.photos/seed/kOfwGr/512/512?blur=1",
                        "format": "decipio",
                        "genre": "coruscus",
                        "publisher": "cupiditas",
                        "createdAt": "2024-03-26T04:06:52.654Z",
                        "updatedAt": "2024-11-14T12:20:12.868Z",
                        "userId": 261
                      },
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaClientKnownRequestError",
                        "code": "P2025",
                        "clientVersion": "5.22.0",
                        "meta": {
                          "modelName": "modelName",
                          "cause": "Record not found."
                        }
                      },
                      "status": 404
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "Delete Book",
        "deprecated": false,
        "description": "",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Book"
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": {
                        "id": 508,
                        "title": "vulnero celebrer addo",
                        "description": "Accusator a enim sunt nam. Damnatio aestus decor ancilla advenio balbus contigo blanditiis votum thorax. Tondeo aeneus vir contego currus quos tabgo.",
                        "cover": "https://picsum.photos/seed/l766Urut/512/512?blur=1",
                        "format": "ventus",
                        "genre": "incidunt",
                        "publisher": "accedo",
                        "createdAt": "2024-04-30T23:35:33.876Z",
                        "updatedAt": "2024-11-13T16:35:54.764Z",
                        "userId": 307
                      },
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaClientKnownRequestError",
                        "code": "P2025",
                        "clientVersion": "5.22.0",
                        "meta": {
                          "modelName": "modelName",
                          "cause": "Record not found."
                        }
                      },
                      "status": 404
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "Update Book",
        "deprecated": false,
        "description": "",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "cover": {
                    "type": "string",
                    "nullable": true
                  },
                  "format": {
                    "type": "string"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "publisher": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Related User"
                  }
                }
              },
              "example": {
                "userId": 1,
                "title": "New Book",
                "description": "Acer comitatus brevis non aegrus. Adopto termes vilis catena.",
                "cover": null,
                "format": "Audiobook",
                "genre": "Business",
                "publisher": "Firebrand Books"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Book"
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": {
                        "id": 197,
                        "title": "illum asporto",
                        "description": "Usitas sodalitas aureus texo summopere dolorem ipsa carbo. Arma adulescens antea aer comitatus. Angustus ab quae animadverto deprecator vorago.",
                        "cover": "https://picsum.photos/seed/yLTRsvg/512/512?blur=1",
                        "format": "deripio",
                        "genre": "ustulo",
                        "publisher": "asporto",
                        "createdAt": "2024-03-08T12:00:56.218Z",
                        "updatedAt": "2024-11-13T16:45:38.990Z",
                        "userId": 68
                      },
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "issues": [
                          {
                            "code": "too_small",
                            "minimum": 2,
                            "type": "string",
                            "inclusive": true,
                            "exact": false,
                            "message": "String must contain at least 2 character(s)",
                            "path": [
                              "title"
                            ]
                          }
                        ],
                        "name": "ZodError"
                      },
                      "status": 400
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaClientKnownRequestError",
                        "code": "P2025",
                        "clientVersion": "5.22.0",
                        "meta": {
                          "modelName": "modelName",
                          "cause": "Record not found."
                        }
                      },
                      "status": 404
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "id",
            "minimum": 1
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "cover": {
            "type": "string",
            "nullable": true
          },
          "format": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "minimum": 1,
            "description": "Related User"
          }
        }
      },
...
    },
    "securitySchemes": false
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Develop Env"
    },
    {
      "url": "https://test.your-api-server.com",
      "description": "Testing Env"
    }
  ],
...
}